+++ R/utilities.R
@@ -79,24 +79,24 @@ assign("prompt_for_download",  TRUE, envir = PublicationBiasBenchmark.private)
   if (is.null(resources)) {
     packageStartupMessage(paste0(
       "This package works with precomputed data, results, and measures.\n",
-      "Specify a location where those resources should be stored and accessed from by using `PublicationBiasBenchmark.options(resources_directory = '/path/')` ", 
-      "or the `PublicationBiasBenchmark_RESOURCES` environmental variable."
-    ))        
+      "Specify a location where those resources should be stored and accessed from by using `PublicationBiasBenchmark.options(resources_directory = '/path/')` ",
+      "or the `PublicationBiasBenchmark_RESOURCES` environment variable."
+    ))
   } else {
     packageStartupMessage(sprintf(paste0(
-      "Data, results, and measures will be stored and accessed from '%1$s'.\n", 
+      "Data, results, and measures will be stored and accessed from '%1$s'.\n",
       "To change the default location, use `PublicationBiasBenchmark.options(resources_directory = '/path/')` ",
-      "or the `PublicationBiasBenchmark_RESOURCES` environmental variable."),
+      "or the `PublicationBiasBenchmark_RESOURCES` environment variable."),
       PublicationBiasBenchmark.private$resources_directory
-    ))    
+    ))
   }
 }
 
 .get_path <- function() {
-  
+
   path <- PublicationBiasBenchmark.get_option("resources_directory")
   if (is.null(path))
     stop("The resources location needs to be specified via the `PublicationBiasBenchmark.options(resources_directory = '/path/')` function.", call. = FALSE)
-  
+
   return(path)
 }